Patch zp/skip-non-stuck-projects to implement waiting routine
Now, if a project is waiting and has a waiting sub-task, it isn’t
stuck. Likewise, if it isn’t a waiting project and it has a waiting
sub-task alongside a next sub-task, it isn’t stuck. However, if none
of those conditions is respected, the project is considered stuck.