org-agenda: Add new group filter to blocks
Also removes the possibility to forward a regex for filtering instead
of a list.  With the new filter taking place in a skip-function rather
than in the agenda-type’s pseudo-regex, it doesn’t make sense to allow
one or the other.  Arguably, it might interesting later down the line
to have another argument to complement the pseudo-regex, but I haven’t
found a use for this yet.