Implement org-capture-web
With org-capture-web.sh, allows grabbing the title and the URL of
webpages and inserting them into an org-capture template.

I had problems getting org-protocol to work on some websites because
of some restrictions.  Whilst org-protocol was cool because it allowed
me to grab the web-page of content of I wanted, I think it’ll already
be super useful to be able to do that for the URL.  No more having to
do copy-pasting between Firefox and Emacs!