agenda: Fix skip functions
When building the agenda, you’ve got a lot of functions working
together.  One of them is the match-function which filters *in* the
potential match based on the regex you provide in tags-todo.  Another
function which is run *after* the match-function is the skip-function
which filters *out* previous matches.

The crucial information which I didn’t have was the following: as soon
as the skip-function finds something to skip (thereby providing a
position from which to resume the search), the match is abandoned and
the match-function is run again after it.

Practically, this means that we don’t need to ensure that every
component in our skip-function is run; we just need to run our
components in a sequence and return as soon as our skip-function
matches something.

I don’t think the order in the sequence is all that important, but as
far as execution speed is concerned, I think I’d rather have the
largest potential skip be the first component to run.