agenda: Optimise agenda with startup-visibility
This is a neat trick.  Looking for solutions to speed up your agenda
always lead you to this one variable, ‘org-agenda-inhibit-startup’
which prevents org-agenda from processing the #+STARTUP visibility
options such as 'folded or 'indent.  This is great for the agenda, but
less so for the visited org-buffer because all the metadatas clutter
the view.

The trick is to expand upon the notion of indirect buffers spawned
from org-agenda, and to consider those the main way to interct with
the mothership, i.e. the org-mode file.  The latter becomes a database
that you’re not supposed to browse, and you access it instead with
org-refile spawning indirect buffers or via the agenda.

Then, those indirect-buffer spawning commands are free to set their
own visibility without changing that of the main buffer.