Allow cloning of buffer in dedicated windows in org-mode
The ‘C-u’ prefixed ‘org-tree-to-indirect-buffer’ didn’t actually work
how I wanted it to work.  It could create a dedicated frame, but it
ended up using it if the command was called again without a ‘C-u’
prefix.  I wanted those dedicated frames to stick rather than the one
spawned before it.

This commit addressed it by nullifying the variable used by
‘org-tree-to-indirect-buffer’ when we make our call, which prevents
the command from killing the previously-spawned buffer.