Implement wrapper for zp/hydra-org-refile
The wrapper allows to reset some toggle-variables when starting a new
hydra-org-refile session (e.g. ‘indirect’ and ‘dedicated’).

I wasn’t able to achieve it with defhydra’s ‘:pre’ and ‘:post’ because
those are run each time a new hydra is created.  In my use-case, since
I’m using multiple levels of hydras, I had to create a wrapper to
encapsulate the session.

‘zp/hydra-org-refile-active’ is a way to handle abnormal exits which
would have caused hydra-org-refile to not reset the variables.