refile: Fix zp/org-refile-from in org-agenda
‘org-agenda-refile’ recenters the agenda window from which it was
called, but ‘zp/org-refile-from failed to preserve the agenda window.

This commit fixes it by encapsulating the call to
‘zp/org-jump-restricted’ within a ‘save-window-excursion’.  However,
I’m making numerous calls to ‘save-window-excursion’, so it might be
a good idea to check whether I can streamline the code.  Still,
I don’t think this is too bad, since each call to an ‘org-refile’
function is done within it’s own ‘save-window-excursion’.