refile: Burry main org buffer after zp/org-refile-main
I’d forgotten t update the code for that command.

It’s a shame that I can’t get this command to use zp/org-refile-to,
and that I have to copy the way the they handle indirect buffer.
I could move it to a function to save some space.  If we think about
it, it’s going to be hard to have some common denominators
between zp/org-refile (used for -main, -other and -restricted), and
zp/org-refile-to (used for -to, and -from).  They do different things
nature, run different checks.

For now, it’ll have to do.