org: Fix advice for APPT_WARNTIME
Now that I’ve got a little more experience with advice, I’ve thought
about a different way to add advice which didn’t cause all the
side-effects when bumping up or down a timestamp without
a APPT_WARNTIME value.

Instead, since I only want this behaviour to happen with interactive
commands requiring a notification, I’ve reused my zp/advise-commands
to add multiple advises and adapted
zp/org-set-appt-warntime-if-timestamp to accept all the arguments
which the interactive command would receive, even though it’s not
doing anything with it.

For more info on advice and arguments:
https://www.gnu.org/software/emacs/manual/html_node/elisp/Advice-combinators.html#Advice-combinators