Fill CREATED with org-agenda sorting function
I’m not sure how I came up with this idea, but it’s bloody brilliant.
I thought it’d feel hackish, but it works great.  If you have
a command that considers every item you have, it’d be a shame not to
have it do some extra legwork, right?