Remove include-groupless-p argument to skip-group
I originally wanted to have it because I didn’t want to lose any task, but
this seems like a really unlikely scenario, and the performance took a bit of
a hit.

Now, if I really want to do that, I can just add ‘+nil’ filter to my ‘Life’
agenda, and it will always catch those items.  I’ve just measured how much
more time it’s taking, and it’s just +150ms with a very basic setup.

Other option would be to check a special agenda matching groupless tasks as
part of my weekly review.

At any rate, there are options, and that setup is still functional if I need
it.