Rework condition for finding next tree after skip
Surprisingly enough, the function worked well without me having to tweak
anything.  I’ve only patched in the function I use to validate my tree, but it
was a happy little accident.

This happened because, at this point in the skip function, we’re telling
org-agenda where to pick up the next search, so the next tree was actually
getting validate on the next skip-function call.  However, since we’re quite
far into the skip-function already, it’s marginally faster to handle the
validation here in a while-loop.