Use copy-tree instead of copy-alist
There was a really subtle problem going on with my org-agenda local-config
where variables which were set to lists would point to the *same* list (i.e.,
the same reference).

The problem was due to my assumption that ‘copy-alist’ was recursive, which it
isn’t.  ‘copy-tree’ is, so changing the function is literally all I needed to
do.