Fix skip function
It turns out that it’s fairly complicated to handle inheritance, and that
I did the wrong way in my agenda-group skip function.

This implementation should hopefully work well.  The conditions speak for
themselves, but put briefly:
- We first check whether any ancestor has the property we seek.
  - If we find one, we can’t reliably tell where the next tree is, so we just go
    to the next headline.
  - If we don’t find one, we can safely search for the next occurrence of the
    property, knowing that if we match something, that tree will be the root
    of that inherited property.

It’s complicated stuff, but I think the condition tree speak for itself.  Took
me a couple of hours to figure out. :D