Tentative change to `declare' to placate checkdoc
I thought this was a solution, but the checkdoc trigger is actually due to
a deeply-nested hydra autogen'd function that has a massive docstring.  I'm
still not sure what check-doc uses to validate if something is a docstring; it
might be the name of the params, or it might be `declare`; I'm more confident
of the latter option.