Uncomment necessary pdf-annot var
I’m pretty sure I’ve disabled this for a compilation error, but I need it
right now and don’t have the time to investigate further.  Plus, there are
already so many problems with this file that I need to address (esp. the
nested hydra long docstrings) that I should relegate any proper fixing to
a full refactoring of my init-files.